Versão alpha 2 release 2.5

+++ database/migrations/2025_03_12_163420_add_description_to_categories_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('categories', function (Blueprint $table) {
-            $table->text('description')->nullable()->after('type');
+            if (!Schema::hasColumn('categories', 'description')) {
+                $table->text('description')->nullable()->after('type');
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('categories', function (Blueprint $table) {
-            $table->dropColumn('description');
+            if (Schema::hasColumn('categories', 'description')) {
+                $table->dropColumn('description');
+            }
         });
     }
 };
